services/glm-ocr-service/requirements.txt
# GLM-OCR Service Requirements

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML dependencies
torch==2.1.0
transformers==4.35.0
pillow==10.1.0

# Utilities
python-multipart==0.0.6
requests==2.31.0

# Monitoring
prometheus-client==0.19.0

# Distributed tracing
jaeger-client==4.8.0
opentracing==2.4.0
